:recycle: Refactor code
Refactor UserType to CustomTypes
